src/requirements.txt
functions-framework==3.*
openai==1.60.1 
python-dotenv==0.19.0